gelf config integrated
