--- gelf/config/Readme.md
@@ -1,3 +1,0 @@
-# Graylog Handler Config
-I put the config in a seperate package in order to be able to import them seperately. 
-Package config introduces a dependency to [github.com/gernoteger/mapstructure-hooks](https://github.com/gernoteger/mapstructure-hooks)
