+++ handler.go
@@ -9,6 +9,8 @@ import (
 	"sync"
 
 	"github.com/go-stack/stack"
+
+
 )
 
 // A Logger prints its log records by writing to a Handler.
